+++ test/obiwan/test.clj
@@ -23,8 +23,10 @@
 (deftest should-mset-and-mget
   (let [ks ["foo" "baz"]
         vs ["bar" "moo"]]
-    (is "OK" (apply redis/mset tt/conn (flatten [ks vs])))
-    (is vs (apply redis/mget tt/conn ks))))
+    (is "OK" (redis/mset tt/conn (->> [ks vs]
+                                      (map vector)
+                                      flatten)))
+    (is vs (redis/mget tt/conn ks))))
 
 (deftest should-incr-and-decr
   (let [k "meaning"
@@ -35,11 +37,23 @@
     (is 0 (redis/decr-by tt/conn k v))
     (is v (redis/incr-by tt/conn k v))))
 
-(deftest should-hmset-and-hgetall
+(deftest should-hset-and-hgetall
   (let [details {"nick" "the blue planet" "age" "4.543 billion years" "mass" "5974000000000000000000000"}
         earth "solar:planet:earth"]
-    (is "OK" (redis/hmset tt/conn earth details))
+    (is "OK" (redis/hset tt/conn earth details))
     (is details (redis/hgetall tt/conn earth))))
+
+(deftest should-run-commands-in-pipeline
+  (let [numbers {"1" "one" "2" "two" "3" "three"}
+        letters {"a" "ey" "b" "bee" "c" "cee"}
+        commands [(redis/hset "numbers" numbers)
+                  (redis/hset "letters" letters)
+                  (redis/hgetall "numbers")
+                  (redis/hgetall "letters")]]
+    (is [(count numbers)
+         (count letters)
+         numbers
+         letters]        (redis/pipeline tt/conn commands))))
 
 (defn run-tests []
   (t/run-all-tests #"obiwan.test.*"))